pyproject.toml
[project]
name = "mcp-trading"
version = "0.1.0"
description = "MCP trading simulation (OpenAI Agents + MCP servers)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=5.22.0",
    "mcp[cli]>=1.5.0",
    "openai-agents>=0.0.15",
    "openai>=1.68.2",
    "pandas>=2.2.0",
    "plotly>=6.0.1",
    "polygon-api-client>=1.14.5",
    "pydantic>=2.11.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
